move SimpleSubstitutionUtils to core
