+++ core/src/main/java/org/vertexium/id/SimpleSubstitutionUtils.java
@@ -1,8 +1,8 @@
-package org.vertexium.accumulo;
+package org.vertexium.id;
 
-import com.beust.jcommander.internal.Lists;
-import com.beust.jcommander.internal.Maps;
 import com.google.common.base.Splitter;
+import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 import org.apache.commons.lang3.tuple.MutablePair;
 import org.apache.commons.lang3.tuple.Pair;
 import org.vertexium.util.IterableUtils;
